Bug fixes and overall improvements all around.

+++ include/sbcqueens-gui/rendering_wrappers/glfw_opengl3_wrapper.h
@@ -176,7 +176,9 @@ namespace ImGUIWrappers {
 				ImGui::SameLine();
 				ImGui::Text("counter = %d", counter);
 
-				ImGui::Text("Application average %.3f ms/frame (%.1f FPS)", 1000.0f / ImGui::GetIO().Framerate, ImGui::GetIO().Framerate);
+				ImGui::Text("Application average %.3f ms/frame (%.1f FPS)",
+					static_cast<double>(1000.0f / ImGui::GetIO().Framerate),
+					static_cast<double>(ImGui::GetIO().Framerate));
 				ImGui::End();
 			}
